requirements.txt
pyaudio
google-cloud-speech
google-cloud-texttospeech
numpy
playsound
nltk
gensim
pymorphy2
